:recycle: Using Strategy patter to avoid DataType enum
Before: Having an enum of possible DataTypes. With that approach, everytime we had to check for an
specific DataType (to decide how to treat it and display it), an if/else statement had to be
created.

Now: Using a class for each possible DataType, with the methods process(), and
display().
* `process()`: This method receives some data (int, array...) and processes it
corresponding to its type. For example, for a `ImagePath` type, this method will receive a path
(str), open it, and return the array that represents the image.
* `display()`: This methods
represents an string representation of the data passed. For example, for a `Categorical` type, this
method will receive an index, and return the string corresponding to the associated category.